+++ Assets/Scripts/architecture_business/use_cases/player_spacecraft/life/SpacecraftPlayerAddLifeUseCase.cs
@@ -3,11 +3,11 @@ using System.Collections.Generic;
 using UnityEngine;
 
 public interface SpacecraftPlayerAddLifeUseCase {
-    public void invoke(int life);
+    public void invoke(float life);
 }
 
 public class SpacecraftPlayerAddLifeUseCaseImpl : SpacecraftPlayerAddLifeUseCase
 {
     private SpacecraftPlayerLifeRepository repo = new SpacecraftPlayerLifeRepositoryImpl();
-    public void invoke(int life) => repo.addLife(life);
+    public void invoke(float life) => repo.addLife(life);
 }
